feat/UT-implement-get-violation-usecase

+++ src/modules/violation/tests/utils/violation/seedViolation.ts
@@ -6,7 +6,7 @@ export const seedViolation = async ({
   id = faker.string.uuid(),
   category = faker.helpers.arrayElement(["A", "B", "C"]),
   violationName = faker.lorem.words(3),
-  penalty = faker.number.float({ min: 100, max: 1000 })
+  penalty = faker.number.int({ min: 100, max: 1000 })
 }: Partial<IViolationSchema>) => {
   return db.violation.create({
     data: {
